chore: rename chloe -> clover, julia -> fish
names are fun. for a few weeks i've been presenting publicly with the
name 'clover', and i like it a lot. for me it's sort of a name tier,
where friends can call me chloe for short, but my full name is clover.
(and for people very close to me, there's more names to unlock).

julia to fish is from a misunderstanding. i used the name julia with her
a lot because it sounded more like a usual name, so i assumed it was.
but no, that's just a nickname. an extra option. her preferred name is
fish. simple as. i am sorry for this confusion.